test(core): consolidate duplicate core suite pairs per #28794
Consolidate five duplicate colocated and __tests__ suites in
packages/core: build-variant, app-registry, outbound-envelope-guard,
callback, and channel-topics. Retain canonical colocated files and
merge all unique assertions from __tests__ before deleting duplicates.

- build-variant: preserve resolution caching (6 tests)
- app-registry: preserve defensive copy, mutation isolation, and
  slug replacement (5 tests)
- outbound-envelope-guard: preserve envelope stream safety and
  reportOutboundEnvelopeBlock bounded preview (11 tests)
- callback: preserve precise 60/61 byte boundary and 15/16 emoji
  unicode byte measurement (18 tests)
- channel-topics: preserve LRU, hydration, persistence, and
  matchTopicRooms cross-channel search (20 tests, 60 total)

Verify: survivor count >= unique(deleted) + unique(survivor) per pair.
All 5 suites pass (60 tests).

Closes #28794
