+++ packages/core/src/build-variant.test.ts
@@ -1,5 +1,7 @@
 /**
  * Unit tests for application build variant resolution (store vs direct).
+ * Consolidated from colocated and __tests__/build-variant suites.
+ * Preserves all unique assertions including resolution caching.
  */
 
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
@@ -61,5 +63,16 @@ describe("build-variant", () => {
 		expect(getDirectDownloadUrl()).toBe("https://eliza.so/download");
 		expect(DEFAULT_BUILD_VARIANT).toBe("direct");
 		expect(BUILD_VARIANTS).toEqual(["store", "direct"]);
+	});
+
+	it("caches the resolved variant across calls", () => {
+		process.env.ELIZA_BUILD_VARIANT = "store";
+		_resetBuildVariantForTests();
+		expect(getBuildVariant()).toBe("store");
+		// After resolution, changing the env has no effect until reset.
+		process.env.ELIZA_BUILD_VARIANT = "direct";
+		expect(getBuildVariant()).toBe("store");
+		_resetBuildVariantForTests();
+		expect(getBuildVariant()).toBe("direct");
 	});
 });
